Refactored XML code -- moved generic stuff under utils/xml/iterparse.py.
Also correctly handles the lifetime of file objects using the `with` statement.
