+++ astropy/io/vo/xmlutil.py
@@ -20,109 +20,9 @@ from .exceptions import (warn_or_raise, vo_warn, vo_raise,
 
 
 __all__ = [
-    'fast_iterparse', 'slow_iterparse', 'get_xml_iterator',
     'check_id', 'fix_id', 'check_token', 'check_mime_content_type',
-    'check_anyuri', 'validate_schema', 'get_xml_encoding',
-    'xml_readlines'
+    'check_anyuri', 'validate_schema'
     ]
-
-
-def fast_iterparse(fd, buffersize=2 ** 10):
-    """
-    Based on :class:`cElementTree.iterparse`, but doesn't ever build a
-    tree at all.  This makes things much faster and more memory
-    efficient.
-
-    Parameters
-    ----------
-    fd : readable file-like object, read function or str path
-
-    Returns
-    -------
-    parts : iterator
-
-        The iterator returns 3-tuples (*event*, *tag*, *data*):
-
-            - *event*: may be either 'start' (when an element begins) or 'end'
-              (when an element is completed).
-
-            - *tag*: The name of the element
-
-            - *data*: Depends on the value of *event*:
-
-                - if *event* == 'start', data is a dictionary of
-                  attributes
-
-                - if *event* == 'end', data is a list of strings, that
-                  when joined is the text content of the element
-    """
-    from xml.parsers import expat
-
-    close_at_end = False
-    if not util.is_callable(fd):
-        if not hasattr(fd, 'read'):
-            close_at_end = True
-            fd = open(fd, 'rb')
-        read = fd.read
-    else:
-        read = fd
-
-    queue = []
-    text = []
-
-    def start(name, attr):
-        queue.append((True, name, attr,
-                      (parser.CurrentLineNumber, parser.CurrentColumnNumber)))
-        del text[:]
-
-    def end(name):
-        queue.append((False, name, ''.join(text).strip(),
-                      (parser.CurrentLineNumber, parser.CurrentColumnNumber)))
-
-    parser = expat.ParserCreate()
-    if not IS_PY3K:
-        parser.returns_unicode = False
-    parser.specified_attributes = True
-    parser.StartElementHandler = start
-    parser.EndElementHandler = end
-    parser.CharacterDataHandler = text.append
-    Parse = parser.Parse
-
-    data = read(buffersize)
-    while data:
-        Parse(data, False)
-        for elem in queue:
-            yield elem
-        del queue[:]
-        data = read(buffersize)
-
-    if close_at_end:
-        fd.close()
-    Parse('', True)
-    for elem in queue:
-        yield elem
-
-# Try to import the C version of the iterparser, otherwise fall back
-# to the Python implementation above.
-slow_iterparse = fast_iterparse
-try:
-    from . import iterparser
-    fast_iterparse = iterparser.IterParser
-except ImportError:
-    pass
-
-
-def get_xml_iterator(source, _debug_python_based_parser=False):
-    """
-    Returns an iterator over the elements of an XML file.  See
-    :class:`fast_iterparse` for more information.
-    """
-    if _debug_python_based_parser:
-        context = slow_iterparse(source)
-    else:
-        context = fast_iterparse(source)
-    iterator = iter(context)
-    return iterator
 
 
 def check_id(ID, name='ID', config={}, pos=None):
@@ -224,46 +124,3 @@ def validate_schema(filename, version='1.2'):
             "data", "VOTable.dtd")
 
     return validate.validate_schema(filename, schema_path)
-
-
-def get_xml_encoding(source):
-    """
-    Determine the encoding of an XML file.  Uses the underlying
-    expat-based iterparser.
-
-    Parameters
-    ----------
-    source : readable file-like object, read function or str path
-
-    Returns
-    -------
-    encoding : str
-    """
-    source = util.convert_to_fd_or_read_function(source)
-    iterator = get_xml_iterator(source)
-    start, tag, data, pos = iterator.next()
-    if not start or tag != 'xml':
-        raise IOError('Invalid XML file')
-
-    return data['encoding']
-
-
-def xml_readlines(source):
-    """
-    Get the lines from a given XML file.  Correctly determines the
-    encoding and always returns unicode.
-
-    Parameters
-    ----------
-    source : readable file-like object, read function or str path
-
-    Returns
-    -------
-    encoding : str
-    """
-    encoding = get_xml_encoding(source)
-
-    with io.open(source, 'rt', encoding=encoding) as input:
-        xml_lines = input.readlines()
-
-    return xml_lines
