+++ astropy/utils/xml/src/.gitignore
@@ -1,0 +1,1 @@
+!*.c
