+++ astropy/utils/xml/src/iterparse.c
@@ -1,54 +1,21 @@
-/*
-Copyright (C) 2008-2010 Association of Universities for Research in Astronomy (AURA)
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are met:
-
-    1. Redistributions of source code must retain the above copyright
-      notice, this list of conditions and the following disclaimer.
-
-    2. Redistributions in binary form must reproduce the above
-      copyright notice, this list of conditions and the following
-      disclaimer in the documentation and/or other materials provided
-      with the distribution.
-
-    3. The name of AURA and its representatives may not be used to
-      endorse or promote products derived from this software without
-      specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY AURA ``AS IS'' AND ANY EXPRESS OR IMPLIED
-WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF
-MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-DISCLAIMED. IN NO EVENT SHALL AURA BE LIABLE FOR ANY DIRECT, INDIRECT,
-INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS
-OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR
-TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE
-USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH
-DAMAGE.
-*/
-
 /******************************************************************************
- * C extension code for vo.table.
+ * C extension code for astropy.utils.xml.iterparse
  *
  * Everything in this file has an alternate Python implementation and
  * is included for performance reasons only.
  *
- * It has three main parts:
+ * It has two main parts:
  *
  *   - An IterParser object which parses an XML file using the expat
  *     library, feeding expat events through a Python iterator.  It is
  *     faster and more memory efficient than the alternatives in the
  *     Python standard library because it does not build a tree of
- *     objects, and also throws away most text nodes, since for VO we
- *     only care about simple text nodes contained between a single
- *     pair of open/close element nodes.  It also has an optimization
- *     for recognizing the most commonly occuring element in a VO
- *     file, "TD".
- *
- *   - An write_tabledata function to quickly write out a Numpy array
- *     in TABLEDATA format.
+ *     objects, and also throws away most text nodes, since for
+ *     astropy.io.vo (the primary user of this library) we only care
+ *     about simple text nodes contained between a single pair of
+ *     open/close element nodes.  It also has an optimization for
+ *     recognizing the most commonly occuring element in a VO file,
+ *     "TD".
  *
  *   - Two functions, escape_xml() and escape_xml_cdata() that escape
  *     XML much faster than the alternatives in the Python standard
@@ -1039,7 +1006,7 @@ static PyTypeObject IterParserType =
 #ifndef IS_PY3K
     0,                          /*ob_size*/
 #endif
-    "IterParser.IterParser",    /*tp_name*/
+    "astropy.utils.xml._iterparser.IterParser",    /*tp_name*/
     sizeof(IterParser),         /*tp_basicsize*/
     0,                          /*tp_itemsize*/
     (destructor)IterParser_dealloc, /*tp_dealloc*/
@@ -1077,300 +1044,6 @@ static PyTypeObject IterParserType =
     0,                          /* tp_alloc */
     IterParser_new,             /* tp_new */
 };
-
-/******************************************************************************
- * Write TABLEDATA
- ******************************************************************************/
-
-#ifdef IS_PY3K
-#  define CHAR Py_UNICODE
-#else
-#  define CHAR char
-#endif
-
-/*
- * Reallocate the write buffer to the requested size
- */
-static int
-_buffer_realloc(
-        CHAR** buffer, Py_ssize_t* buffer_size, CHAR** x, Py_ssize_t req_size)
-{
-    Py_ssize_t  n       = req_size;
-    CHAR *      new_mem = NULL;
-
-    if (req_size < *buffer_size) {
-        return 0;
-    }
-
-    /* Calculate the next-highest power of two */
-    n = next_power_of_2(n);
-
-    if (n < req_size) {
-        PyErr_SetString(PyExc_MemoryError, "Out of memory for XML text.");
-        return -1;
-    }
-
-    new_mem = realloc((void *)*buffer, n * sizeof(CHAR));
-    if (new_mem == NULL) {
-        PyErr_SetString(PyExc_MemoryError, "Out of memory for XML text.");
-        return -1;
-    }
-
-    *x = (CHAR *)new_mem + (*x - *buffer);
-    *buffer = new_mem;
-    *buffer_size = n;
-
-    return 0;
-}
-
-/*
- * Write *indent* spaces to the buffer
- */
-static int
-_write_indent(CHAR** buffer, Py_ssize_t* buffer_size,
-              CHAR** x, Py_ssize_t indent)
-{
-    if (_buffer_realloc(buffer, buffer_size, x,
-                        (*x - *buffer + indent))) {
-        return 1;
-    }
-
-    for (; indent; --indent) {
-        *(*x)++ = ' ';
-    }
-
-    return 0;
-}
-
-/*
- * Write a string into a buffer.  On Python 3 the string and buffer
- * are multibyte
- */
-static int
-_write_string(CHAR** buffer, Py_ssize_t* buffer_size,
-              CHAR** x, const CHAR* src, const Py_ssize_t len) {
-    if (_buffer_realloc(buffer, buffer_size, x,
-                        (*x - *buffer + len))) {
-        return 1;
-    }
-
-    while (*src != (CHAR)0) {
-        *(*x)++ = *src++;
-    }
-
-    return 0;
-}
-
-/*
- * Write an 8-bit C string to a possibly Unicode string
- */
-static int
-_write_cstring(CHAR** buffer, Py_ssize_t* buffer_size,
-               CHAR** x, const char* src, const Py_ssize_t len) {
-    if (_buffer_realloc(buffer, buffer_size, x,
-                        (*x - *buffer + len))) {
-        return 1;
-    }
-
-    while (*src != (char)0) {
-        *(*x)++ = *src++;
-    }
-
-    return 0;
-}
-
-/*
- * Write a TABLEDATA element tree to the given write method.
- *
- * The Python arguments are:
- *
- * *write_method* (callable): A Python callable that takes a string
- *    (8-bit on Python 2, Unicode on Python 3) and writes it to a file
- *    or buffer.
- *
- * *array* (numpy structured array): A Numpy record array containing
- *    the data
- *
- * *mask* (numpy array): A Numpy array which is True everywhere a
- *    value is missing.  Must have the same shape as *array*.
- *
- * *converters* (list of callables): A sequence of methods which
- *    convert from the native data types in the columns of *array* to
- *    a string in VOTABLE XML format.  Must have the same length as
- *    the number of columns in *array*.
- *
- * *write_null_values* (boolean): When True, write null values in
- *    their entirety in the table.  When False, just write empty <TD/>
- *    elements when the data is null or missing.
- *
- * *indent* (integer): The number of spaces to indent the table.
- *
- * *buf_size* (integer): The size of the write buffer.
- *
- * Returns None.
- */
-static PyObject*
-write_tabledata(PyObject* self, PyObject *args, PyObject *kwds)
-{
-    /* Inputs */
-    PyObject* write_method = NULL;
-    PyObject* array = NULL;
-    PyObject* mask = NULL;
-    PyObject* converters = NULL;
-    int write_null_values = 0;
-    Py_ssize_t indent = 0;
-    Py_ssize_t buf_size = (Py_ssize_t)1 << 8;
-
-    /* Output buffer */
-    CHAR* buf = NULL;
-    CHAR* x;
-
-    Py_ssize_t nrows = 0;
-    Py_ssize_t ncols = 0;
-    Py_ssize_t i, j;
-    int write_full;
-    int all;
-    PyObject* numpy_module = NULL;
-    PyObject* numpy_all_method = NULL;
-    PyObject* array_row = NULL;
-    PyObject* mask_row = NULL;
-    PyObject* array_val = NULL;
-    PyObject* mask_val = NULL;
-    PyObject* converter = NULL;
-    PyObject* all_masked_obj = NULL;
-    PyObject* str_val = NULL;
-    PyObject* tmp = NULL;
-    CHAR* str_tmp = NULL;
-    Py_ssize_t str_len = 0;
-    PyObject* result = 0;
-
-    if (!PyArg_ParseTuple(args, "OOOOinn:write_tabledata",
-                          &write_method, &array, &mask, &converters,
-                          &write_null_values, &indent, &buf_size)) {
-        goto exit;
-    }
-
-    if (!PyCallable_Check(write_method)) goto exit;
-    if (!PySequence_Check(array)) goto exit;
-    if (!PySequence_Check(mask)) goto exit;
-    if (!PyList_Check(converters)) goto exit;
-    indent = CLAMP(indent, (Py_ssize_t)0, (Py_ssize_t)80);
-    buf_size = CLAMP(buf_size, (Py_ssize_t)1 << 8, (Py_ssize_t)1 << 24);
-
-    if ((numpy_module = PyImport_ImportModule("numpy")) == NULL) goto exit;
-    if ((numpy_all_method = PyObject_GetAttrString(numpy_module, "all"))
-        == NULL) goto exit;
-
-    if ((nrows = PySequence_Size(array)) == -1) goto exit;
-    if ((ncols = PyList_Size(converters)) == -1) goto exit;
-
-    if ((buf = malloc((size_t)buf_size * sizeof(CHAR))) == NULL) goto exit;
-
-    for (i = 0; i < nrows; ++i) {
-        if ((array_row = PySequence_GetItem(array, i)) == NULL) goto exit;
-        if ((mask_row = PySequence_GetItem(mask, i)) == NULL) goto exit;
-
-        x = buf;
-        if (_write_indent(&buf, &buf_size, &x, indent)) goto exit;
-        if (_write_cstring(&buf, &buf_size, &x, " <TR>\n", 6)) goto exit;
-
-        for (j = 0; j < ncols; ++j) {
-            if ((converter = PyList_GET_ITEM(converters, j)) == NULL) goto exit;
-            if ((array_val = PySequence_GetItem(array_row, j)) == NULL) goto exit;
-            if ((mask_val = PySequence_GetItem(mask_row, j)) == NULL) goto exit;
-
-            if (write_null_values) {
-                write_full = 1;
-            } else {
-                if (mask_val == Py_False) {
-                    write_full = 1;
-                } else if (mask_val == Py_True) {
-                    write_full = 0;
-                } else {
-                    if ((all_masked_obj =
-                         PyObject_CallFunctionObjArgs(numpy_all_method, mask_val, NULL))
-                        == NULL) goto exit;
-                    if ((all = PyObject_IsTrue(all_masked_obj)) == -1) {
-                        Py_DECREF(all_masked_obj);
-                        goto exit;
-                    }
-                    Py_DECREF(all_masked_obj);
-
-                    write_full = !all;
-                }
-            }
-
-            if (write_full) {
-                if (_write_indent(&buf, &buf_size, &x, indent)) goto exit;
-                if (_write_cstring(&buf, &buf_size, &x, "  <TD>", 6)) goto exit;
-
-                if ((str_val =
-                     PyObject_CallFunctionObjArgs(converter, array_val, mask_val, NULL))
-                    == NULL) goto exit;
-#ifdef IS_PY3K
-                if ((str_tmp = PyUnicode_AsUnicode(str_val)) == NULL) {
-                    Py_DECREF(str_val);
-                    goto exit;
-                }
-                str_len = PyUnicode_GetSize(str_val);
-#else
-                if (PyString_AsStringAndSize(str_val, &str_tmp, &str_len)
-                    == -1) {
-                    Py_DECREF(str_val);
-                    goto exit;
-                }
-#endif
-                if (_write_string(&buf, &buf_size, &x, str_tmp, str_len)) {
-                    Py_DECREF(str_val);
-                    goto exit;
-                }
-
-                Py_DECREF(str_val);
-
-                if (_write_cstring(&buf, &buf_size, &x, "</TD>\n", 6)) goto exit;
-            } else {
-                if (_write_indent(&buf, &buf_size, &x, indent)) goto exit;
-                if (_write_cstring(&buf, &buf_size, &x, "  <TD/>\n", 8)) goto exit;
-            }
-
-            Py_DECREF(array_val); array_val = NULL;
-            Py_DECREF(mask_val);  mask_val = NULL;
-        }
-
-        Py_DECREF(array_row); array_row = NULL;
-        Py_DECREF(mask_row);  mask_row = NULL;
-
-        if (_write_indent(&buf, &buf_size, &x, indent)) goto exit;
-        if (_write_cstring(&buf, &buf_size, &x, " </TR>\n", 7)) goto exit;
-
-        /* NULL-terminate the string */
-        *x = (CHAR)0;
-#ifdef IS_PY3K
-        if ((tmp = PyObject_CallFunction(write_method, "u#", buf, x - buf))
-            == NULL) goto exit;
-#else
-        if ((tmp = PyObject_CallFunction(write_method, "s#", buf, x - buf))
-            == NULL) goto exit;
-#endif
-        Py_DECREF(tmp);
-    }
-
-    Py_INCREF(Py_None);
-    result = Py_None;
-
- exit:
-    Py_XDECREF(numpy_module);
-    Py_XDECREF(numpy_all_method);
-
-    Py_XDECREF(array_row);
-    Py_XDECREF(mask_row);
-    Py_XDECREF(array_val);
-    Py_XDECREF(mask_val);
-
-    free(buf);
-
-    return result;
-}
 
 /******************************************************************************
  * XML escaping
@@ -1545,8 +1218,6 @@ escape_xml_cdata(PyObject* self, PyObject *args, PyObject *kwds)
 
 static PyMethodDef module_methods[] =
 {
-    {"write_tabledata", (PyCFunction)write_tabledata, METH_VARARGS,
-     "Fast C method to write tabledata"},
     {"escape_xml", (PyCFunction)escape_xml, METH_VARARGS,
      "Fast method to escape XML strings"},
     {"escape_xml_cdata", (PyCFunction)escape_xml_cdata, METH_VARARGS,
@@ -1571,7 +1242,7 @@ static int module_clear(PyObject* m)
 
 static struct PyModuleDef moduledef = {
     PyModuleDef_HEAD_INIT,
-    "iterparser",
+    "_iterparser",
     "Fast XML parser",
     sizeof(struct module_state),
     module_methods,
@@ -1584,7 +1255,7 @@ static struct PyModuleDef moduledef = {
 #  define INITERROR return NULL
 
 PyMODINIT_FUNC
-PyInit_iterparser(void)
+PyInit__iterparser(void)
 #else /* Not PY3K */
 #  define INITERROR return
 
@@ -1593,7 +1264,7 @@ PyInit_iterparser(void)
 #  endif
 
 PyMODINIT_FUNC
-inititerparser(void)
+init_iterparser(void)
 #endif
 {
     PyObject* m;
@@ -1601,7 +1272,7 @@ inititerparser(void)
 #ifdef IS_PY3K
     m = PyModule_Create(&moduledef);
 #else
-    m = Py_InitModule3("iterparser", module_methods, "Fast XML parser");
+    m = Py_InitModule3("_iterparser", module_methods, "Fast XML parser");
 #endif
 
     if (PyType_Ready(&IterParserType) < 0)
